requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
webauthn==2.2.0
PyJWT==2.9.0
